requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.3.3
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.5
distlib==0.3.8
distro==1.9.0
Django==5.0.3
django-compat==1.0.15
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
drf-standardized-errors==0.13.0
drf-yasg==1.21.7
exceptiongroup==1.2.1
filelock==3.13.1
firebase-admin==6.4.0
Flask==3.0.3
frozenlist==1.4.1
google-api-core==2.17.1
google-api-python-client==2.121.0
google-apps-meet==0.1.6
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.6
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
lambda-pkg-resources==0.0.6
MarkupSafe==2.1.5
MouseInfo==0.1.3
msgpack==1.0.8
multidict==6.0.5
oauthlib==3.2.2
openai==1.13.3
packaging==23.2
pdfkit==1.0.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodome==3.20.0
pycryptodome-test-vectors==1.0.14
pycryptodomex==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.1
pywhatkit==5.4
PyYAML==6.0.1
redis==5.0.2
referencing==0.33.0
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rpds-py==0.18.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
virtualenv==20.25.1
wcwidth==0.2.13
Werkzeug==3.0.3
wikipedia==1.4.0
yarl==1.9.4
